Improved closing logic for emoji panel and emoji suggestion panel
