+++ Explorer/Assets/DCL/Chat/ChatController.cs
@@ -16,6 +16,7 @@ using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using System.Threading;
 using UnityEngine;
+using UnityEngine.EventSystems;
 using UnityEngine.InputSystem;
 using Utility;
 
@@ -44,16 +45,17 @@ namespace DCL.Chat
         private readonly IEventSystem eventSystem;
         private readonly World world;
         private readonly Entity playerEntity;
+        private readonly Mouse device;
+        private readonly DCLInput dclInput;
+        private readonly ChatCommandsHandler commandsHandler;
 
         private CancellationTokenSource cts;
         private CancellationTokenSource emojiPanelCts;
         private SingleInstanceEntity cameraEntity;
-        private readonly DCLInput dclInput;
-
-        private readonly ChatCommandsHandler commandsHandler;
         private (IChatCommand command, Match param) chatCommand;
         private CancellationTokenSource commandCts = new ();
         private bool isChatClosed = false;
+        private IReadOnlyList<RaycastResult> raycastResults;
 
         public override CanvasOrdering.SortingLayer Layer => CanvasOrdering.SortingLayer.Persistent;
 
@@ -92,6 +94,7 @@ namespace DCL.Chat
             // Adding two elements to count as top and bottom padding
             chatMessages.Add(new ChatMessage(true));
             chatMessages.Add(new ChatMessage(true));
+            device = InputSystem.GetDevice<Mouse>();
         }
 
         protected override void OnViewInstantiated()
@@ -107,9 +110,8 @@ namespace DCL.Chat
             viewInstance.InputField.onSubmit.AddListener(OnSubmit);
             viewInstance.CloseChatButton.onClick.AddListener(CloseChat);
             viewInstance.LoopList.InitListView(0, OnGetItemByIndex);
-            emojiPanelController = new EmojiPanelController(viewInstance.EmojiPanel, emojiPanelConfiguration, emojiMappingJson, emojiSectionViewPrefab, emojiButtonPrefab, dclInput, eventSystem);
+            emojiPanelController = new EmojiPanelController(viewInstance.EmojiPanel, emojiPanelConfiguration, emojiMappingJson, emojiSectionViewPrefab, emojiButtonPrefab);
             emojiPanelController.OnEmojiSelected += AddEmojiToInput;
-            emojiPanelController.OnPanelClosedByClick += () => viewInstance.EmojiPanelButton.SetState(false);
 
             emojiSuggestionPanelController = new EmojiSuggestionPanel(viewInstance.EmojiSuggestionPanel, emojiSuggestionViewPrefab, dclInput);
             emojiSuggestionPanelController.OnEmojiSelected += AddEmojiFromSuggestion;
@@ -120,6 +122,34 @@ namespace DCL.Chat
             viewInstance.ChatBubblesToggle.Toggle.SetIsOnWithoutNotify(nametagsData.showChatBubbles);
             dclInput.UI.Submit.performed += OnSubmitAction;
             OnToggleChatBubblesValueChanged(nametagsData.showChatBubbles);
+        }
+
+        protected override void OnViewShow()
+        {
+            base.OnViewShow();
+            dclInput.UI.Click.performed += OnClick;
+        }
+
+        protected override void OnViewClose()
+        {
+            base.OnViewClose();
+            dclInput.UI.Click.performed -= OnClick;
+        }
+
+        private void OnClick(InputAction.CallbackContext obj)
+        {
+            raycastResults = eventSystem.RaycastAll(device.position.value);
+            var clickedOnPanel = false;
+            foreach (RaycastResult raycasted in raycastResults)
+                if (raycasted.gameObject == viewInstance.EmojiPanel.gameObject || raycasted.gameObject == viewInstance.EmojiSuggestionPanel.gameObject)
+                    clickedOnPanel = true;
+
+            if (!clickedOnPanel)
+            {
+                viewInstance.EmojiPanelButton.SetState(false);
+                viewInstance.EmojiPanel.gameObject.SetActive(false);
+                emojiSuggestionPanelController!.SetPanelVisibility(false);
+            }
         }
 
         private void OnChatViewPointerExit() =>
